Remove DataController, TExtField and Move TagCloud & ToggleButton to List components
