--- packages/web/examples/TextField/.babelrc
@@ -1,14 +1,0 @@
-{
-    "presets": [
-        [
-            "env", {
-                "targets": {
-                    "browsers": ["last 2 versions", "safari >= 7"]
-                },
-                "modules": false,
-                "loose": true
-            }
-        ],
-        "react"
-    ]
-}
