--- packages/web/examples/TextField/.gitignore
@@ -1,8 +1,0 @@
-# Build
-dist
-
-# Packages
-node_modules
-
-# Directory
-.directory
