--- packages/web/examples/TextField/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<title>TextField</title>
-	<link href="https://use.fontawesome.com/releases/v5.0.2/css/all.css" rel="stylesheet">
-</head>
-
-<body>
-	<div id="root">
-	</div>
-	<script src="dist/index.js">
-	</script>
-</body>
-
-</html>
