--- packages/web/examples/TextField/src/index.css
@@ -1,148 +1,0 @@
-/* layout */
-
-body {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding-top: 15px;
-}
-
-.row {
-	display: flex;
-	flex-direction: row;
-	width: 100%;
-}
-
-.col {
-    flex: 1;
-	padding: 15px;
-}
-
-.row > .col:first-child {
-	border-right: 1px solid #ccc;
-	max-width: 400px;
-}
-
-.row > .col:last-child {
-	background: #fafafa;
-}
-
-.flex {
-	display: flex;
-}
-
-.wrap {
-	flex-wrap: wrap;
-}
-
-.column {
-	flex-direction: column;
-}
-
-.align-center {
-	align-items: center;
-}
-
-.justify-center {
-	justify-content: center;
-}
-
-.justify-space-between {
-	justify-content: space-between;
-}
-
-.text-center {
-	text-align: center;
-}
-
-@media all and (max-width: 767px) {
-    .row {
-        flex-direction: column;
-    }
-
-    .row > .col:first-child {
-        border-right: none;
-        max-width: none;
-    }
-}
-
-/* apps */
-
-.authors-list {
-	color: #9d9d9d;
-	font-weight: bold;
-}
-
-.ratings-list {
-	padding: 10px 0;
-}
-
-.avg-rating {
-	color: #6b6b6b;
-	margin-left: 5px;
-}
-
-.stars {
-	color: gold;
-}
-
-.location {
-	color: salmon;
-	margin-right: 5px;
-}
-
-.meetup-location {
-	margin: 4px 0;
-}
-
-.book-title {
-	white-space: normal;
-	margin-top: 4px;
-}
-
-.book-title-card {
-	white-space: normal;
-	margin-top: 4px;
-	max-height: 45px;
-}
-
-.book-image {
-	height: 150px;
-	width: 110px;
-	background-size: cover;
-}
-
-.book-header {
-	font-weight: bold;
-	margin-bottom: 5px;
-}
-
-.book-content {
-	background: white;
-	margin: 10px 0;
-	box-shadow: 0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24);
-}
-
-.meetup-title {
-	white-space: normal;
-}
-
-.meetup-topics {
-	height: 35px;
-	overflow: hidden;
-}
-
-.meetup-topic {
-	background-color: #dedede;
-    color: #555;
-	padding: 5px 10px;
-	margin: 5px;
-	border-radius: 4px;
-}
-
-.meetup-topic:first-child {
-	margin-left: 0;
-}
-
-.col .meetup-list-image {
-	background-size: cover;
-}
