+++ packages/web/src/components/list/TagCloud.js
@@ -27,7 +27,7 @@ import Title from '../../styles/Title';
 import TagList from '../../styles/TagList';
 import Container from '../../styles/Container';
 import { connect, isFunction } from '../../utils';
-import { getAggsQuery } from '../list/utils';
+import { getAggsQuery } from './utils';
 
 class TagCloud extends Component {
 	constructor(props) {
