+++ packages/web/src/index.js
@@ -1,12 +1,8 @@
 import '@appbaseio/reactivecore/lib/utils/polyfills';
 
 import ReactiveBase from './components/basic/ReactiveBase';
-import TextField from './components/basic/TextField';
 import SelectedFilters from './components/basic/SelectedFilters';
-import ToggleButton from './components/basic/ToggleButton';
 import NumberBox from './components/basic/NumberBox';
-import DataController from './components/basic/DataController';
-import TagCloud from './components/basic/TagCloud';
 import ReactiveComponent from './components/basic/ReactiveComponent';
 
 import DatePicker from './components/date/DatePicker';
@@ -18,6 +14,8 @@ import SingleDropdownList from './components/list/SingleDropdownList';
 import MultiDropdownList from './components/list/MultiDropdownList';
 import SingleDataList from './components/list/SingleDataList';
 import MultiDataList from './components/list/MultiDataList';
+import TagCloud from './components/list/TagCloud';
+import ToggleButton from './components/list/ToggleButton';
 
 import SingleRange from './components/range/SingleRange';
 import MultiRange from './components/range/MultiRange';
@@ -38,11 +36,9 @@ import ResultList from './components/result/ResultList';
 export {
 	// basic
 	ReactiveBase,
-	TextField,
 	SelectedFilters,
 	ToggleButton,
 	NumberBox,
-	DataController,
 	TagCloud,
 	ReactiveComponent,
 	// date
